backend/requirements.txt
fastapi
uvicorn
tortoise-orm
aerich
pydantic
pydantic-settings
python-dotenv
aiomysql
asyncmy
cryptography
